move store into /pkg/
